pyproject.toml
[project]
name = "voice-app"
version = "0.1.0"
description = "Voice-to-voice AI app: Mic → STT → LLM → TTS → Speaker"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "python-dotenv>=1.0",
    "faster-whisper>=1.0",
    "huggingface-hub>=1.7.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
voice-app = "voice_app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
